Add VTS test for VNDK ABI
Add VtsVndkAbiTest that compares dump files with libraries on device.
The dump includes symbols and vtables of all shared libraries marked as
vendor_available.

For each dump file, the test case first searches /vendor/lib[64], and
then /system/lib[64] for the library. If the library is not found, the
comparison is skipped.

Test: vts-tradefed run commandAndExit vts -m VtsVndkAbi
Bug: 37488823
Change-Id: I97503ffb255ea939605984508b4a2f82ea0a8409