Update VndkAbiDump.proto to include symbols
AbiDump message now contains two fields: vtables and symbols. "symbols"
stores the dynamic symbols of the library.

VTableEntry message now records the kind of the entry and "is_inlined"
field is removed because it's not working as expected.

Symbol message format is added to record the dynamic symbols of a
library. Symbol name and symbol binding (global or weak) is recorded.

Bug: 80217511
Test: python -c 'import VndkAbiDump_pb2'
Change-Id: Icf2ccaa6ed6e1bf21ad729a5a6e23dc941dd2ad3