Let dump_abi.py read JSON format dump
ReadLsdump() now reads lsdump contents in json or protobuf text format.
It first tries to decode the contents of file in json format.
If json decoding fails then it tries to decode it in protobuf text
format. If both decoding fails, it raises an exception supplying the
reason why both decoding fails.

Bug: 80217511
Bug: 112877706
Test: # generate json format dump then do
Test: ./dump_abi.py ../../../../build/make/target/product/vndk/current.txt
Change-Id: Iea69e167bbcfb155c6b34b8f79bfcd90503b6b52